UI run-through, component and style rationalisation
Full app works as a local player on a single book, next step will be to
export the state to the backend and allow for multiple devices/players.

Slider re-worked so that it is rendered once and then slides in and out
of view. This has the added benefit that the slide out animation logic
can live where it belongs in the slider but the business of
opening/closing the sliders belongs to the parent. Probably need to pull
the common slider props into an interface but will wait till I add the
popups as I feel like the modal components will have the same sort of
properties so we could use the same interface across all the components
with modes.

Switched to sass to give us variables and extracted the common colours
to the new colors module.

A bit of general touching up and pulling common css styles up into the
App.scss file so that controls are consistent across the application. We
decided that the palette and thickness controls are special cases of
buttons so these are not bound by the same styling.

Add a bunch of packages I'm sure I didn't end up using. Will need to
review which dependencies are actually required before production build
but npm's package model is confusing for me so for now I will just leave
them be as they don't seem to be causing problems.